moved authorizations and validations into controller

+++ src/controllers/comment.js
@@ -1,10 +1,13 @@
 import { sendDataResponse } from '../utils/responses.js'
+import User from '../domain/user.js'
 import {
   create,
   getAllForPost,
   updateComment,
-  deleteComment
+  deleteComment,
+  getCommentById
 } from '../domain/comment.js'
+import { findById } from '../domain/post.js'
 import { Prisma } from '@prisma/client'
 
 export const createComment = async (req, res) => {
@@ -51,7 +54,7 @@ export const getAllComments = async (req, res) => {
 }
 
 export const editComment = async (req, res) => {
-  const id = Number(req.params.commentid)
+  const id = Number(req.params.commentId)
 
   const { content } = req.body
   try {
@@ -76,17 +79,33 @@ export const editComment = async (req, res) => {
 }
 
 export const deleteCommentFromPost = async (req, res) => {
-  const id = Number(req.params.commentid)
+  const id = Number(req.params.commentId)
+  const postId = Number(req.params.id)
   try {
+    const post = await findById(postId)
+    if (!post) {
+      return sendDataResponse(res, 404, { error: 'post not found' })
+    }
+    const comment = await getCommentById(id)
+    console.log(comment.user)
+    if (!comment) {
+      return sendDataResponse(res, 404, { error: 'comment not found' })
+    }
+    if (req.user.id !== comment.userId || req.user.role !== 'TEACHER') {
+      return sendDataResponse(res, 403, {
+        authorization: 'You are not authorized to perform this action'
+      })
+    }
     const deletedComment = await deleteComment(id)
-
+    const author = await User.findById(deletedComment.userId)
+    delete author.passwordHash
     const deletedCommentWithAuthor = {
       comment: {
         id: deletedComment.id,
         content: deletedComment.content,
         createdAt: deletedComment.createdAt,
         updatedAt: deletedComment.updatedAt,
-        author: { ...req.user }
+        author: { ...author }
       }
     }
     return sendDataResponse(res, 200, { deletedCommentWithAuthor })
